Remove copy pasted code from json_pointer_error Rename json_pointer_error.h json_pointer_exception
